requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.12
fastmcp==2.3.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
markdown-it-py==3.0.0
mcp==1.9.0
mdurl==0.1.2
openapi-pydantic==0.5.1
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
typer==0.15.4
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
websockets==15.0.1
